refactor: unify synthetic runner execution runtime
